refactor: extract provekit-field-bn254; make common field-agnostic via a registered provider
Replace the in-common ProofField trait-impl with a runtime FieldHashProvider
registry: common now reaches the field-native Merkle engine ids, public-input
binding hashes, and Fiat-Shamir sponge through a registered provider, and names
no field crate (skyscraper/poseidon2/ntt) — verified via cargo tree.

- New provekit-field-bn254 crate: holds the moved skyscraper/poseidon2/ntt
  modules, implements FieldHashProvider for bn254, and exposes register() which
  installs the NTT + Merkle hash engines and the provider (same pattern as
  whir's ENGINES/NTT registries).
- common drops its ntt/poseidon2/skyscraper dependencies and the provekit_ntt
  feature; the orphan rule keeps the per-field impl out of common, so the
  registry is the seam that yields true crate-level field isolation (a bn254
  binary never links a goldilocks crate) and keeps the follow-up additive.
- Registration moves to the callers: prover/verifier register at entry, the
  r1cs-compiler frontend registers before reading the engine id at scheme
  construction, and the FFI registers in pk_init (all transitional until A6).
- Public-input binding KATs relocate to provekit-field-bn254 (they exercise the
  real registered provider; a dev-only dependency cycle would split common into
  two instances with separate registries).

bn254 stays byte-identical; full workspace builds, KATs and an end-to-end Noir
prove/verify pass.
