+++ provekit/common/Cargo.toml
@@ -11,14 +11,8 @@ repository.workspace = true
 [features]
 default = ["parallel"]
 parallel = []
-provekit_ntt = []
 
 [dependencies]
-# Workspace crates
-ntt.workspace = true
-poseidon2.workspace = true
-skyscraper.workspace = true
-
 # Noir language
 acir.workspace = true
 noirc_abi.workspace = true
@@ -62,6 +56,10 @@ ignored = ["keccak"] # Intentionally anchored to keep Noir beta.19-compatible RC
 
 [dev-dependencies]
 divan.workspace = true
+# Dev-only cycle: the bn254 field crate depends on common, but common's tests
+# and benches need a registered field provider / the bn254 NTT. Cargo permits
+# dev-dependency cycles (they are excluded from the normal build graph).
+provekit-field-bn254.workspace = true
 
 [target.'cfg(not(target_arch = "wasm32"))'.dev-dependencies]
 proptest.workspace = true
