+++ provekit/field/bn254/src/poseidon2/sponge.rs
@@ -30,9 +30,9 @@
 //! `Poseidon2::hash()`); this is the duplex transcript sponge.
 
 use {
-    crate::utils::{bytes_to_field, field_to_bytes_le},
     ark_bn254::Fr,
     poseidon2::permutation::poseidon2_permutation,
+    provekit_common::utils::{bytes_to_field, field_to_bytes_le},
     spongefish::{DuplexSponge, Permutation},
 };
