+++ provekit/prover/src/lib.rs
@@ -104,7 +104,7 @@ impl Prove for NoirProver {
         self,
         acir_witness_idx_to_value_map: WitnessMap<NoirElement>,
     ) -> Result<NoirProof> {
-        provekit_common::register_ntt();
+        provekit_field_bn254::register();
 
         let mut public_input_indices = self.program.functions[0].public_inputs().indices();
         public_input_indices.sort_unstable();
@@ -275,7 +275,7 @@ impl Prove for NoirProver {
 impl Prove for MavrosProver {
     #[cfg(feature = "witness-generation")]
     fn prove(mut self, input_map: InputMap) -> Result<NoirProof> {
-        provekit_common::register_ntt();
+        provekit_field_bn254::register();
 
         let params = crate::input_utils::ordered_params_from_btreemap(&self.abi, &input_map)?;
         let phase1 = mavros_interpreter::run_phase1(
