+++ provekit/r1cs-compiler/Cargo.toml
@@ -11,6 +11,9 @@ repository.workspace = true
 [dependencies]
 # Workspace crates
 provekit-common.workspace = true
+# Frontend registers the bn254 field backend before building/using a scheme
+# (scheme construction reads the field-native Merkle engine id).
+provekit-field-bn254.workspace = true
 poseidon2.workspace = true
 
 # Noir language
