+++ provekit/r1cs-compiler/src/whir_r1cs.rs
@@ -55,6 +55,9 @@ impl WhirR1CSSchemeBuilder for WhirR1CSScheme {
         has_public_inputs: bool,
         hash_config: HashConfig,
     ) -> Self {
+        // Scheme construction reads the field-native Merkle engine id, so the
+        // bn254 field backend must be registered first.
+        provekit_field_bn254::register();
         assert_eq!(
             num_challenges,
             challenge_offsets.len(),
@@ -153,6 +156,9 @@ impl WhirR1CSSchemeBuilder for WhirR1CSScheme {
         has_public_inputs: bool,
         hash_config: HashConfig,
     ) -> Self {
+        // Scheme construction reads the field-native Merkle engine id, so the
+        // bn254 field backend must be registered first.
+        provekit_field_bn254::register();
         debug_assert_eq!(
             num_challenges,
             challenge_offsets.len(),
